Add gify test
